Extract MovieRating to own file.
